arm64: dts: qcom: kona-usb: Fix interrupts and qserdes conf for USB
The usb configuration was "suboptimal": the lane mode wasn't correctly
set and so the gains.
This was actually producing no issue but lots of confusion.

Most importantly, the interrupt configuration was seriously wrong,
with firing the DP HSPHY interrupt on both levels for literally no
reason (when it has to be triggered only when rising, as that's when
we need to take action)... and that was the same for dm, as well..

What a mess...