msm: synx: Prevent synx base ids from being negative
Left-shift on a tgid can yield negative values when tgid is large. This
is especially probable with pid namespaces in place (e.g. when using
Waydroid). This will result in failed synx handle allocation and may
confuse userspace HALs.

We will just make sure that the first bit of the base id can never be 1.
This will certainly result in collisions, but it should be fine in all
but extremely unlikely circumstances where two synx clients happen to
share the same lower 15 bits in their tgid -- way less common than
having a large tgid that results in negative values when left-shifted by
16 bits.